switch where functions live
