+++ tests/unit/commands/test_test.py
@@ -4,6 +4,7 @@ from ...testcases import DustyTestCase
 from ..utils import get_app_dusty_schema, get_lib_dusty_schema
 from dusty.commands import test
 from dusty.schemas.base_schema_class import DustySpecs
+from dusty.source import Repo
 
 @patch('dusty.commands.test.initialize_docker_vm')
 @patch('dusty.commands.test.get_docker_client')
@@ -125,3 +126,40 @@ class TestTestsCommands(DustyTestCase):
 
         fake_update_test.assert_has_calls([])
         fake_make.assert_has_calls([call('app1', fake_specs)])
+
+class TestGetDependentRepos(DustyTestCase):
+    @patch('dusty.schemas.base_schema_class.get_specs_from_path')
+    def test_get_all_repos_for_app_or_library_app(self, fake_get_specs):
+        fake_get_specs.return_value = {'apps': {'app1': get_app_dusty_schema(
+                                                    {'depends': {'apps': ['app2', 'app3'],
+                                                                 'libs': ['lib1']},
+                                                     'repo': '/gc/app1'}),
+                                                'app2': get_app_dusty_schema(
+                                                    {'depends': {'apps': ['app4'],
+                                                     'libs': []},
+                                                     'repo': '/gc/app2'}),
+                                                'app3': get_app_dusty_schema({'depends': {'apps': [], 'libs': []}, 'repo': '/gc/app3'}),
+                                                'app4': get_app_dusty_schema({'depends': {'apps': [], 'libs': []}, 'repo': '/gc/app4'})},
+                                       'libs': {'lib1': get_lib_dusty_schema({'depends': {'libs': ['lib2']}, 'repo': '/gc/lib1'}),
+                                                'lib2': get_lib_dusty_schema({'depends': {'libs': []}, 'repo': '/gc/lib2'})}}
+
+        self.assertEquals(set(test._get_all_repos_for_app_or_library('app1')),
+                          set([Repo('/gc/app1'), Repo('/gc/lib1'), Repo('/gc/lib2')]))
+
+    @patch('dusty.schemas.base_schema_class.get_specs_from_path')
+    def test_get_all_repos_for_app_or_library_lib(self, fake_get_specs):
+        fake_get_specs.return_value = {'apps': {'app1': get_app_dusty_schema(
+                                                    {'depends': {'apps': ['app2', 'app3'],
+                                                                 'libs': ['lib1']},
+                                                     'repo': '/gc/app1'}),
+                                                'app2': get_app_dusty_schema(
+                                                    {'depends': {'apps': ['app4'],
+                                                     'libs': []},
+                                                     'repo': '/gc/app2'}),
+                                                'app3': get_app_dusty_schema({'depends': {'apps': [], 'libs': []}, 'repo': '/gc/app3'}),
+                                                'app4': get_app_dusty_schema({'depends': {'apps': [], 'libs': []}, 'repo': '/gc/app4'})},
+                                       'libs': {'lib1': get_lib_dusty_schema({'depends': {'libs': ['lib2']}, 'repo': '/gc/lib1'}),
+                                                'lib2': get_lib_dusty_schema({'depends': {'libs': []}, 'repo': '/gc/lib2'})}}
+
+        self.assertEquals(set(test._get_all_repos_for_app_or_library('lib1')),
+                          set([Repo('/gc/lib1'), Repo('/gc/lib2')]))
